SDA-4757 - Fix config race condition
